+++ 03week/guiPigLatin.css
@@ -3,7 +3,7 @@ body {
   min-height: 760px;
   background: #000;
   mix-blend-mode: color #000;
-  background-image: url(./img/angry-birds-alone-pig-hd-wallpaper-wallpaper-preview.jpg);
+  /* background-image: url(./img/angry-birds-alone-pig-hd-wallpaper-wallpaper-preview.jpg); */
   background-repeat: no-repeat;
   /* background-position: center; */
   background-position: fixed;
@@ -11,21 +11,41 @@ body {
   padding-bottom: 10px;
 }
 
+#myVideo {
+  position: fixed;
+  right: 0;
+  bottom: 0;
+  min-width: 100%;
+  min-height: 100%;
+}
+
+/* Add some content at the bottom of the video/page */
+.content {
+  position: fixed;
+  bottom: 0;
+  /* background: rgba(0, 0, 0, 0.5); */
+  /* color: #f1f1f1; */
+  width: 100%;
+  height: 720px;
+  padding: 0px;
+}
+
 header {
-  /* margin: 0 auto; */
-  padding-left: 150px;
-  padding-top: 130px;
+  margin: 0 auto;
+  /* padding-left: 150px; */
+  /* padding-top: 10px; */
   height: 207px;
-  /* text-align: center; */
-  color: rgba(71, 173, 68, 0.829);
+  text-align: center;
+  color: rgba(240, 241, 240, 0.829);
   font-size: 70px;
   /* font-family: "Courier New", Courier, monospace; */
   font-weight: lighter;
-  letter-spacing: 12px;
+  letter-spacing: 15px;
 }
 
 h1 {
-  color: rgba(71, 173, 68, 0.829);
+  margin-left: 595px;
+  color: rgba(240, 241, 240, 0.829);
   font-size: 20px;
   padding-top: 20px;
   padding-left: 150px;
@@ -36,21 +56,21 @@ h1 {
 
 #translation {
   padding-top: 20px;
-  color: rgba(71, 173, 68, 0.829);
+  color: rgba(240, 241, 240, 0.829);
 }
 
 button {
-  margin-left: 150px;
-  color: rgba(71, 173, 68, 0.829);
+  margin-left: 740px;
+  color: rgba(240, 241, 240, 0.829);
   /* color: rgba(94, 85, 126, 0.781); */
   font-size: 20px;
   /* font-family: "Courier New", Courier, monospace; */
   /* font-weight: lighter; */
   background-color: transparent;
-  border-top-color: rgba(105, 90, 156, 0.24);
-  border-left-color: rgba(105, 90, 156, 0.24);
-  border-right-color: rgba(105, 90, 156, 0.781);
-  border-bottom-color: rgba(105, 90, 156, 0.781);
+  border-top-color: rgba(240, 241, 240, 0.397);
+  border-left-color: rgba(240, 241, 240, 0.397);
+  border-right-color: rgba(240, 241, 240, 0.829);
+  border-bottom-color: rgba(240, 241, 240, 0.829);
   /* border-left-width: 1.5px;
   border-right-width: 1.5px;
   border-right-width: 2px;
@@ -61,23 +81,24 @@ button {
 }
 
 span {
+  margin-left: 750px;
   /* color: rgba(71, 173, 68, 0.829); */
-  color: rgba(105, 90, 156, 0.781);
-  font-size: 60px;
+  color: rgb(0, 0, 0);
+  font-size: 80px;
   background-color: transparent;
-  padding-left: 150px;
+  /* padding-left: 150px; */
   /* font-family: "Courier New", Courier, monospace; */
-  font-weight: lighter;
+  /* font-weight: lighter; */
   letter-spacing: 5px;
 }
 
 #word {
-  color: rgba(71, 173, 68, 0.829);
+  color: rgba(240, 241, 240, 0.829);
   /* color: rgba(94, 85, 126, 0.781); */
   font-size: 20px;
   padding-left: 10px;
   background-color: transparent;
-  border-color: rgba(105, 90, 156, 0.781);
+  border-color: rgba(240, 241, 240, 0.829);
   border-radius: 3%;
   border-width: 1.25px;
   /* font-family: "Courier New", Courier, monospace; */
