+++ 03week/guiPigLatin.html
@@ -8,18 +8,75 @@
     <title>Document</title>
   </head>
   <body>
-    <header>pigLatin</header>
-    <br />
-    <h1>Word: <input id="word" /></h1>
-    <br />
-    <button onclick="handleClick()">
-      Oinkyay!
-    </button>
-    <br />
-    <h1 id="translation">
-      Translation:
-    </h1>
-    <span id="translateWord">...</span>
+    <!-- <video id="myVideo">
+      <source
+        src="./img/Baby goat meets Pulp Fiction - say what one more time!.mp4"
+        type="video/mp4"
+      />
+    </video> -->
+    <video id="myVideo">
+      <source src="./img/what.mp4" type="video/mp4" />
+    </video>
+    <div class="content">
+      <header>pigLatin</header>
+      <br />
+      <h1>Word: <input id="word" /></h1>
+      <br />
+      <!-- <button onclick="playVid()" type="button">Play Video</button>
+    <button onclick="pauseVid()" type="button">Pause Video</button><br /> -->
+      <button onclick="playVid(), handleClick()">
+        Oinkyay!
+      </button>
+      <br />
+      <h1 id="translation">
+        Translation:
+      </h1>
+      <span id="translateWord">...</span>
+    </div>
+    <!-- <div id="player"></div>
+
+    <script>
+      // 2. This code loads the IFrame Player API code asynchronously.
+      var tag = document.createElement("script");
+
+      tag.src = "https://www.youtube.com/iframe_api";
+      var firstScriptTag = document.getElementsByTagName("script")[0];
+      firstScriptTag.parentNode.insertBefore(tag, firstScriptTag);
+
+      // 3. This function creates an <iframe> (and YouTube player)
+      //    after the API code downloads.
+      var player;
+      function onYouTubeIframeAPIReady() {
+        player = new YT.Player("player", {
+          height: "390",
+          width: "640",
+          videoId: "FcoQXC23krI",
+          events: {
+            onReady: onPlayerReady,
+            onStateChange: onPlayerStateChange,
+          },
+        });
+      }
+
+      // 4. The API will call this function when the video player is ready.
+      function onPlayerReady(event) {
+        event.target.playVideo();
+      }
+
+      // 5. The API calls this function when the player's state changes.
+      //    The function indicates that when playing a video (state=1),
+      //    the player should play for six seconds and then stop.
+      var done = false;
+      function onPlayerStateChange(event) {
+        if (event.data == YT.PlayerState.PLAYING && !done) {
+          setTimeout(stopVideo, 6000);
+          done = true;
+        }
+      }
+      function stopVideo() {
+        player.stopVideo();
+      }
+    </script> -->
   </body>
-  <script src="guiPigLatin_redo.js"></script>
+  <script src="guiPigLatin.js"></script>
 </html>
