--- 03week/guiPigLatin_redo.js
@@ -1,77 +1,0 @@
-"use strict";
-
-console.log("loading js script");
-
-// +++++++++++ handleClick function ++++++++++++++
-
-function handleClick() {
-  let audio = new Audio("audio/snortingPig.mp3");
-  audio.play();
-
-  var myInput = document.getElementById("word");
-  let word = myInput.value;
-  console.log("word :", word);
-
-  let translation = pigLatin(word);
-  console.log("translation", translation);
-
-  let translateWord = document.getElementById("translateWord");
-  translateWord.innerText = translation;
-}
-
-// +++++++++++ pigLatin function ++++++++++++++
-
-// Find First Vowel
-
-function findFirstVowel(word) {
-  const vowel = ["a", "e", "i", "o", "u"];
-  console.log("=====Find First Vowel=====");
-
-  for (let j = 0; j < word.length; j++) {
-    var firstVowel = vowel.indexOf(word[j]);
-
-    console.log(firstVowel);
-
-    if (vowel.indexOf(word[j]) !== -1) {
-      return j;
-    }
-  }
-}
-
-// Translate
-
-function pigLatin(word) {
-  word = word.toLowerCase();
-  let firstVowel = findFirstVowel(word);
-
-  if (firstVowel > 0) {
-    return word.substring(firstVowel) + word.substring(0, firstVowel) + "ay";
-  } else {
-    return word + "yay";
-  }
-}
-
-// Tests
-
-if (typeof describe === "function") {
-  describe("#pigLatin()", () => {
-    it("should translate a simple word", () => {
-      assert.equal(pigLatin("car"), "arcay");
-      assert.equal(pigLatin("dog"), "ogday");
-    });
-    it("should translate a complex word", () => {
-      assert.equal(pigLatin("create"), "eatecray");
-      assert.equal(pigLatin("valley"), "alleyvay");
-    });
-    it('should attach "yay" if word begins with vowel', () => {
-      assert.equal(pigLatin("egg"), "eggyay");
-      assert.equal(pigLatin("emission"), "emissionyay");
-    });
-    it("should lowercase and trim word before translation", () => {
-      assert.equal(pigLatin("HeLlO "), "ellohay");
-      assert.equal(pigLatin(" RoCkEt"), "ocketray");
-    });
-  });
-} else {
-  getPrompt();
-}
